Rename the SDK to just iotconnect.c and h
